# api/swagger.yaml
definitions:
  go-platform_internal_models_dogs.DogImageResponse:
    properties:
      breed:
        type: string
      image_url:
        type: string
    type: object
  go-platform_pkg_utils_http-utils.ErrorDetail:
    properties:
      field:
        type: string
      message:
        type: string
    type: object
  go-platform_pkg_utils_http-utils.ErrorResponse:
    properties:
      details:
        items:
          $ref: '#/definitions/go-platform_pkg_utils_http-utils.ErrorDetail'
        type: array
      log_timestamp:
        type: string
      message:
        type: string
      request_id:
        type: string
      status:
        type: integer
    type: object
info:
  contact: {}
  description: Go Platform API
  title: Go Platform
  version: "1.0"
paths:
  /api/v1/dogs/{breed}/image:
    get:
      description: Retrieves a random dog image for the specified breed, downloads
        it, and uploads to S3
      parameters:
      - description: Dog breed
        in: path
        name: breed
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: S3 URL of the uploaded image
          schema:
            $ref: '#/definitions/go-platform_internal_models_dogs.DogImageResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/go-platform_pkg_utils_http-utils.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/go-platform_pkg_utils_http-utils.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/go-platform_pkg_utils_http-utils.ErrorResponse'
      summary: Get random dog image by breed
      tags:
      - Dogs
  /live:
    get:
      consumes:
      - application/json
      produces:
      - application/json
      responses:
        "200":
          description: ok
          schema:
            type: string
        "500":
          description: internal server error
          schema:
            type: string
      summary: Health check
      tags:
      - Health
swagger: "2.0"